Removed reference to '/usr/lib64/c++init.o' from the default compile command.
That file may not be installed in that location (especially if Unix
modules are in use with multiple compiler versions) and should be
found automatically by the compiler.

If it is not found, that is more likely a problem with the compiler
installation or less likely with the user's environment settings.

Spencer Swift