Broken periodic behavior for longitude.
Issue: The specification of longitude on a spherical grid not unique. For example the
longitude of one grid can be expressed as ranging from -180 to 180 degrees while the
second from 0 to 360 degrees. The current search stage of the remapping routines does
not know how to reconcile these two equivalent periodic ranges.

Solution: The solution is to shift one or more of the coordinate systems to some agreed
upon range. We select 0 to 360. The natural place to perform this shift is within the
ESMF_RegridStore. The RegridStore calls the RegridCreate which then calls routines
specific to the particular regridding method, such as conservative, bilinear, nearest
neighbor, etc. The natural place for the shift is within these routines.

The local variables dstLocalCoordX, ... , dstLocalCornerY are pointers to the original
grid specification using an ESMF_InternArrayGetData. The pointers to the destination
grid information srcGatheredCoordX, ... , srcGatheredCornerY come from the calls
ESMF_RouteRunF90PtrR821D.

Because the local coordinate varables are pointers, changing the local variables also
changes the original grid. Since we don't want to change what the user chose, we
change the calls starting at line 330 for associating the destination grid with the
pointers dstLocalCoordX, ... , dstLocalCornerY from ESMF_DATA_REF to ESMF_DATA_COPY.