Support for getting number of acc devices via OpenCL
The ESMF application should now be able to get the number of
accelerator devices using ESMF_VMGet(). This change includes
support for getting the number of acc devices via the OpenCL
framework. The change also includes changes to the build/configure
system so that user can build ESMF with OpenCL framework support.