Initial cut of the 1d array manipulation algorithms
This commit adds utility algorithms that can be used to create
and manipulate PET lists. All of the algorithms work on one
dimensional arrays and can be used to simplify creation of
user petlists.

* Adding Value Iterators : Value iterator is a data structure
that can represent ranges of values (start, end, stride) and
user-defined custom set of values (values specified via a buffer
filled by the user). The value iterator can be used to iterate
over the range of values specified by the user.

* Adding algorithms that can manipulate value iterators. In this
commit an algorithm to split a value iterator using a color array
is included.

* Adding algorithms to split and join buffers (arrays) using
value iterators. There are also algorithms to create (destroy)
and fill buffers using value iterators.

* The algorithms were implemented in C++ and the necessary
fortran 2003 bindings (Fortran to C) to access these algorithms
are also included.