Modifications for the PGI compiler
PGI compilers don't allow (bug?) passing a procedure pointer
to a subroutine that accepts a subroutine as the argument. The
other compilers (GNU, Intel and NAG) allow this behavior.

So to get the code to work with PGI we accept a procedure(...),
instead of a subroutine, as the argument to the function. This
works on all compilers.