Adding an simple OpenCL matrix mult for acc comp
The accelerated component now performs multiplication of two
matrices using OpenCL on the accelerator device. Adjusting
the size of the matrices multiplied should give us enough
work for the GPU/device.

A serial version of the matrix multiplication code is also
included (used to verify the result).

The OpenCL matrix multiplication code is simplistic and can
use a lot of improvements (to improve performance)