Using max (not sum) to find number devs on a node
The accelerator devices are available to all MPI ranks on a node
or a single rank has access to all devices on the node. So using
MAX intead of SUM (adding up all devices accessible from procs
on the same node) to find number of devices available on a node.