Adding a two concurrent comp test for Mapper
Adding a test with two concurrent components and a coupler.

This test is a copy of the ESMF_ConcurrentComponent test, but
has been modified so that it runs on any number of PETs. Also
the distribution of PETs across two components is changed from
a static allocation to dynamic. Now the first component gets
the first half of PETs and the second compoent gets the rest of
the PETs, the coupler as before runs on all PETs.