Adding a simple two component test for mapper
Adding a simple test for mapper with two concurrent components.
Both components have the same clock settings and use half of
the PETs (each).

The mapper interfaces are invoked to initialize and optimize
the PET lists but the components are not recreated (as of this
commit) using optimized PET lists. The test is currently used for
testing the mapper interfaces (rather than the mapper
functionality).