Adding a basic mapper optimize function
Adding a basic optimize function for the mapper. This function
partition pets based on the execution time (elapsed wallclock
time) for each component.

The algorithm first does a bottom up parse of the execution
block tree to update the execution times for each execution block.
Then the algorithm proceeds to partition pets parsing the
execution block tree from the top.