Remove states from the application
Removing the import and export states from the components. This
should result in a bare bones application that would be a better
candidate for testing recreation of user components.

(Note: As of this commit running a component after recreating
it does not work, even without any import or export states. So
for now we just print out the optimized pet list from the mapper)