Remove dups in dependency graph
Since we parse for two kinds of dependencies (a comp phase depends
on any phase of the same comp before, data dep based on a connector
before the current comp phase) its possible that we add the
same node twice as the child of a node.

This commit checks whether a node is already added as a child before
adding it (don't add a dup child).

Also adding the line number from the NUOPC run sequence to the
tokenized run sequence. This helps us in debugging.