Enhance 2d poly fit to nd poly fit
Allow creation of n degree polynomial fits for the user data.
The previous version restricted the number of degrees to 2, now
the degree of the polynomial is accepted as an input argument.

Also adding unit tests for 2d, 3d and 4d fit polynomials for the
same data.

Also modifying the rest of the utils to use the new poly fit
interface.