Fix the check for wgtd avg vs solver
If the number of stored vals is less than 3 we use the weighted
average and once we have at least 3 vals (wallclock times) we
switch to the solver (and try to fit a poly).

The previous code incorrectly checked the number of components
to switch between weighted avg and the solver.