Add option to scale and center polynomial vars
Adding an option to scale and center polynomial variables. The
polynomial fit algorithm now scales and centers the input data
(x values) such that the values are centered around the mean
and have a unit standard deviation. This scaling and centering
of the data improves the stability of the algorithms that uses
the fit polynomial (stability of the solver)

* A new class to represent scaling and centering of polynomial
  variables is added
* The fit algorithm now scales and centers the input data before
  finding a fit polynomial. This scaling and centering info is
  saved in the polynomial and is used during evaluation (to scale
  and center the user variable value before evaluation)
* Without scaling and centering the values the solver returned
  -ve values for PETs when used from the Load balancer.