Adding a poly class with dep functions
Adding a polynomial class to represent polynomials with variables
that have dependent variables. The relationship between the
polynomial variables and the dependent variables are provided
as a vector of multivariate linear polynomials.

Also adding a unit test

As of this commit the functionality related to dependent variables
are not implemented/used/tested.