Add a dampening const and niters to solver
Adding a dampening constant to the solver to reduce the size of
steps (corrective steps) in the Newton Raphson solver. The
dampening constant is currently set to 0.01

Also adding a dampening constant to the load balancer, used for
equations passed to the solver. Currently its set to 1.0 (no
dampening)

The solver in the load balancer is now run iteratively 100 times
(setting the number of max iters in the solver), instead of
relying on the defaults in the solver