Update comp info in loadbalancer with user info
Update the information cached in the load balancer about PET
layouts based on updated information from the user.

The load balancer predicts the wallclock time for a layout and
provides the layout to the user. Once the user runs the app
using the layout and sets the measured wallclock time , we now
update the cached (predicted) wallclock time for the layout. We
also measure and store the error (for debugging purposes) in
the wallclock time prediction.