Support for reshaping and scaling solution
Adding support for reshaping solutions in the solver. Also
adding support for scaling and centering the constraint functions
used by the solver.

Reshaping the solution, re-assigns the Xi from the solver so that
Sum(Xi) is equal to the sum of the initial values of Xis. It also
makes sure that Xi does not go into the negative range.

Scaling and centering the solution makes sure that all the
constraint functions have similar weight. The coefficients of the
constraint functions are scaled such that the results of the
constraint functions with the initial values of Xis are centered
around the mean and have a standard deviation of 1. This change
makes sure that the individual constraint functions have similar
weights.