A simple execution simulator
Adding a simple execution simulator for a component layout. Once
the component info objects are updated with a pet layout, the
execution simulator uses information about the wallclock execution
times for each component provided to it to "execute" the
components.

The execution simulator can be used to update the timings
(wallclock timings) for a list of components and to find the
wallclock time of an application