Get correct PET layout in load balancer
In the load balancer we now find the PET layout based on the
pet ranges of the different components. This change is required
to represent components that run sequentially with other
components. The old code assumed that all components were
running concurrently to each other.

Also a new function is added to add backup information about
component layouts (instead of directly updating the backup
info collection)

* Adding util functions to get pet layout and wallclock time
  for an optimized PET layout