Splitting backup info into candidates and actual
Information on component layouts are either generated by the
mapper or provided by the user. The ones provided by the user
are now in a separate collection and is used to find the final
optimal PET layout. The potential candidate layouts are stored
separately and is used to provide the user with the next
potential candidate layout.