Adding the basics of F2C interface for Mapper
Adding the basic structure and some Fortran to C interfaces for
the Mapper. Also adding the C interfaces for the Mapper.

The appropriate directories and source/include files required
for the Fortran to C bindings is added.

Currently only create and destroy methods are implemented.

The test case with two concurrent components is modified
accordingly to use the new Fortran interfaces for the
Mapper. The test case fails for now (since not all interfaces
are implemented), but builds and runs with the new interface.