Adding a util func to approx fit polynomials
This new utility function with perform an approximate polynomial
fit for a given polynomial using the range of x values provided
by the user and some auto generated x values.

The auto generated x values are exponentially increasing to cover
the lower range. A second set of auto generated x values are
normally distributed around the mean of the user provided x values.

A unit test is also added to test the function.