Start generalizing constructPioDecomp to handle multiple tiles
For now it still just handles a single tile, but I have put in place
some generality to start moving towards handling multiple tiles. And it
correctly handles the decomposition for a single tile of a multi-tile
array now, which it didn't do before.

Note that I changed the calculation of arrayShape in
PIO_IODescHandler::constructPioDecomp: I'm pretty sure it is incorrect
to have tile in here: this is documented as having dimension
[redDimCount*ssiLocalDeCount]. So in principle we should have the DE
number in here, but we can leave that out as long as we assume no more
than one DE per PE.