Remove some unnecessary isOpen checks
These checks are either (1) redundant with existing checks in the
IO_Handler functions called or (2) unnecessary because the IO_Handler
functions gracefully handle the case of a closed file. And these checks
are going to be hard to maintain with an upcoming change to make isOpen
work on just one tile.