Darwin.gfortranclang: always use clang++ as a linker
Exception handling only works if we link with clang++ rather than
gfortran, even if the main program is in Fortran.

Also set -fno-backtrace; this is required to avoid a hang when trying to
print a backtrace.

There are still a lot of test failures, but this is a little better than
before.

Note that the hard-coded addition of Fortran-specific MPI libraries is
temporary; I want to make this more dynamic.