Get Fortran OpenMPI libraries dynamically
Before the changes on this branch (i.e., before changing the link step
to be done with clang instead of gfortran), the libs.openmpif90 script
was used. However, that is designed for using a Fortran MPI
wrapper (explicitly linking against the c++ mpi library in that case);
now we want to do the reverse.