Rework logic in nfconfigtest
On my Mac with homebrew, "nf-config --version" gives
"4.6.0-development", which led the old logic in nfconfigtest to think
that it wasn't working, since it was looking for "netCDF-Fortran" in the
output.

The new logic checks more specifically for the failure mode that
initially led to the creation of
nfconfigtest (https://sourceforge.net/p/esmf/tix/3614881/).

Resolves esmf-org/esmf#81