Darwin gfortranclang: Use C++ linker even for C programs
This is needed to access functions from the C++ standard library;
otherwise, we get link errors when building with
ESMF_SHARED_LIB_BUILD=OFF. We could alternatively specify -lstdc++, and
that seems to solve the problem in at least some cases, but this may
lead to problems such as those with exception handling that we
encountered when linking with gfortran, which led us to use the C++
linker when linking Fortran programs. So this use of the C++ linker when
linking C programs is analogous to the use of the C++ linker when
linking Fortran programs with this compiler combination.

See https://github.com/esmf-org/esmf/issues/112 for details.