Change order of checks for file type auto-detection
The key part of this is putting the check for lon and lat last, so that
this is only triggered if one of the other file types (ESMFMESH or
SCRIP) isn't detected - since other file types besides GRIDSPEC can
potentially have lon and lat variables with these units.

I am also putting the check for ESMFMESH-style variables before
SCRIP-style variables, with the rationale that the former seems slightly
more specific than the latter, and I want to go from most to least
specific in these checks.

Resolves esmf-org/esmf#94