Respond to an ESMPY_DATA_NEW_DIR environment variable
The behavior is now:
- As before, if you set ESMPY_DATA_DIR, the assumption is that this
directory is already populated with the necessary data, so no automatic
downloads are attempted.
- The new behavior is: otherwise, if you set ESMPY_DATA_NEW_DIR, data
will be automatically downloaded to the path set by this variable
- Otherwise, as before, data will be automatically downloaded to a data
directory within the esmpy installation

See https://github.com/conda-forge/esmpy-feedstock/issues/78,
https://github.com/conda-forge/esmpy-feedstock/pull/79 and
https://github.com/esmf-org/esmf-support/issues/355 for motivation.

This commit also introduces a do_download function to replace some
duplicate checks of ESMPY_DATA_DIR, so that the various logic for the
data dir can stay co-located.