Change values in unit tests of I/O with replicated dims
Add more variability to the arrays.

This started as an attempt to fix what appear to be messed up values in
some of the I/O of Arrays with replicated dimensions, especially this
one:

netcdf Array_repli1_first {
dimensions:
	myData_dim001 = 10 ;
	myData_dim002 = 3 ;
variables:
	double myData(myData_dim002, myData_dim001) ;
data:

 myData =
  0, _, 12345, _, 24690, _, 37035, _, 49380, 61725,
  0, _, 12345, _, 24690, _, 37035, _, 49380, 61725,
  0, _, 12345, _, 24690, _, 37035, _, 49380, 61725 ;
}

But this pattern of missing data persists no matter how I fill the
arrays. So I'm giving up on trying to fix this pattern of missing data,
but am leaving in place this slight improvement in the setting of
values (so that there's more variability from one point to another).