Remove a seemingly-unnecessary error check
This error check was added in commit
7379b259f2d2d621e335204026b2e2644318772c (by Gerhard, Jan 12, 2018):
"Add checks to prevent redist_arraycreate1de() from trying to redist an
array where the index space is too small to span all PETs."

Note that there is no error check like this in the case where we're not
doing a redist, which makes me think that this may not be a fundamental
PIO limitation. Furthermore, in a test program, it worked to remove this
error check when writing and reading a 2x4 Field with 10 PETs (and going
through the redist function). So it seems like this check isn't
necessary.