Add an env var to allow skipping performance tests
Add a new environment variable ESMF_TESTPERFORMANCE. If set to "OFF",
this will skip checking performance in performance tests.

This is useful on machines where performance tests are highly variable
and can lead to spurious failures. e.g., this happens on green (a MacOS
/ Darwin system), where the nightly cron job can run very slowly,
leading to failures in some performance tests (generally in
ESMF_ArrayRedistPerfUTest).