Fix ESMPy logic for determining if PIO is enabled
This no longer depends on whether we're using mpiuni

I thought that we might need to add some logic to handle the case where
the user explicitly sets ESMF_PIO=OFF: I thought that would appear in
the esmf.mk file. But it turns out that ESMF_PIO doesn't appear in the
esmf.mk file in that case, so the logic here seems correct.