Remove a limitation in mpiuni's MPI_Type_create_indexed_block
I had a check that the displacements go in increments of blocklength,
but from some experimentation with checking the MPI_Type_size of the new
type returned from this function from openmpi, it seems that the
resulting type's size is independent of the displacements. So I don't
think we need this check.