Change ESMPy locstream creation methods to avoid duplicate points
The existence of duplicate points in the example locstream creation
methods resulted in unpredictable results that differed between
machines. This commit changes the locstream spherical creation methods
so that all points are unique, which should result in consistent
behavior across machines.

This also changes the tolerance for the affected tests so that the new
tolerances are in line with the actual relative errors. I determined the
new values by running the tests; the actual relative errors were:
- For test_grid_locstream_regrid, 4.80e-6
- For test_locstream_grid_regrid, 3.07e-7

Resolves esmf-org/esmf#227