Fix sprintf issues
Fixes uses of sprintf where the output string is also used in the input,
which can result in garbled time strings.

Also changes these sprintf calls to snprintf to avoid possible buffer
overflows, and checks these calls to ensure that the output variable is
large enough to hold the relevant string, returning an error code if
not.

Resolves esmf-org/esmf#284