Fix LocStream's __getitem__ for python >= 3.12
The logic in LocStream's `__getitem__` was built around the assumption
that `slice` objects aren't hashable. However, starting with Python
3.12, `slice` objects *are* hashable. This commit changes the
implementation of this logic while keeping the same result as before for
both slices and strings.

Note that, in theory, there are a number of other types that can be
returned from `get_formatted_slice`, depending on the type of the `slc`
argument to `__getitem__`: a list of multiple `slice(None)` objects, a
`np.ndarray`, a tuple of slices, or a tuple of `np.ndarray`s. These
cases all would have led to errors with the previous logic (entering the
`except` block but then hitting an error in `slc_ls.stop`, since none of
those types have a `stop` attribute), and now will lead to different
errors (in trying to execute `__getitem__` on these objects).

See https://github.com/orgs/esmf-org/discussions/313 for some context.

Co-authored-by: Xylar Asay-Davis <xylarstorm@gmail.com>