Fix creation of mask in grid_create functions
The previous logic (involving the use of `.any()`) was causing an error
in recent versions of NumPy. On inspection, the use of `.any()` looked
wrong, and not what was intended. In addition, the use of `.data` in
other functions caused errors; I think these functions that used `.data`
were never being called, so these errors weren't showing up. This commit
fixes the logic to what I think was intended.

This fix caused a failure in `test_field_regrid_zeroregion`. Bob Oehmke
and I looked at the test together and felt that the test was incorrect
and should actually be asserting that values are -100 where the mask is
0 (rather than having 0 values there), so I have made this change as
well. This was passing previously because the mask was never 0 due to
the above bug.

Resolves https://github.com/esmf-org/esmf/issues/329