ESMPy: provide supported versions in pyproject.toml
Provide earliest supported version for python - bump from 3.7 to 3.8
because that is the earliest we're testing with. (Not listing a latest
supported version, as per advice in
https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#requires-python-upper-bounds
and
https://packaging.python.org/en/latest/guides/dropping-older-python-versions/#specify-the-version-ranges-for-supported-python-distributions).

Provide earliest and latest supported versions for numpy. The earliest
version is based on the earliest version we're testing with; the latest
version is based on presumed backwards incompatibility with a major
version bump.