Declare some arguments to f_esmf_regridstorefile as optional
Mostly stay consistent with f_esmf_regridstore in this respect, except:

(1) I'm leaving routehandle non-optional because making it optional
would require some extra logic (interacting with createRoutehandle), and
things seem to be working as is.

(2) I'm also leaving rc non-optional because making it optional would
require some extra logic.

(Note that it seems potentially problematic that these are both optional
in f_esmf_regridstore, but I want to get out of this rabbit hole, so I'm
leaving things as they are there for now.)

Resolves https://github.com/esmf-org/esmf/issues/407