Don't try to support ESMC_GridGetLocalDECount for ProtoGrid
The Grid Create/Set/Commit pattern doesn't appear to be supported yet
via the C API, so there didn't seem to be an easy way to test the
retrieval via the ProtoGrid. It seemed best to remove this untested
complexity and just support getting the localDECount for a fully-formed
Grid.