Get cubed_sphere_to_mesh_regrid example working with any PET count
Previously, before the work on this branch to enable multiple DEs per
PET in ESMPy, this example was partially working with 6 PETs: it ran to
completion, but had large errors because it wasn't working correctly.
(This is because it was set up to use multiple DEs per PET even with 6
PETs; it worked correctly if the decomposition was changed so that it
used just 6 DEs total.) With fewer than 6 PETs, it was very broken.

Now it works and gives the expected low errors with any number of PETs.