Fix ESMPy Grid class to work with 0 DEs per PET
These changes are needed when we have a cubed-sphere Grid that has more
than 6 PETs but a number of PETs that is not a multiple of 6: in this
case, some PETs have 0 DEs, and thus attempting to access data for DE #0
(the first DE in python indexing) results in an out-of-bounds error. So
this commit reworks some code to avoid this issue.