Allow tests in test_field.py to pass with larger PET counts
Before this change, at least one test in test_field.py failed with 12
processors (they passed with 10; I didn't try 11). (The same failure
happened from ESMF v8.8.1 - so this was not a new problem due to the
recent changes to enable multiple DEs per PET in ESMPy.) I think this is
because the test involved a 10x10 grid, and so with > 10 processors, we
got an out-of-bounds error in trying to set mask values. This commit
adds a guard conditional to prevent this issue.

After this change, all tests in test_field.py, test_grid.py and
cubed_sphere_to_mesh_regrid.py pass with 8, 10, 12, 16 and 18
processors, in addition to the standard tests on 1, 4 and 6 processors.