Use scipy's netcdf instead of netCDF4 for testing netcdf files
I had problems with netCDF4 on my Mac: see
https://github.com/esmf-org/esmf-test-scripts/issues/102. scipy's netcdf
shouldn't give these kinds of problems because it doesn't depend on
having a netcdf or hdf5 library. (However, one downside is that it can't
read netcdf4/hdf5-formatted files.)

Note that, in contrast to the earlier code, I am *requiring* this
package, rather than silently moving on if the netcdf package can't be
found. This is because I want to know if some tests aren't being run.